break out the car, deck and player classes to a core module for reuse
