+++ core/src/main/java/org/gardler/biglittlechallenge/core/App.java
@@ -1,0 +1,13 @@
+package org.gardler.biglittlechallenge.core;
+
+/**
+ * Hello world!
+ *
+ */
+public class App 
+{
+    public static void main( String[] args )
+    {
+        System.out.println( "Hello World!" );
+    }
+}
