+++ core/src/main/java/org/gardler/biglittlechallenge/core/model/Player.java
@@ -1,0 +1,28 @@
+package org.gardler.biglittlechallenge.core.model;
+
+public class Player {
+
+	String name;
+	Deck deck;
+	
+	public Deck getDeck() {
+		return deck;
+	}
+
+	public void setDeck(Deck deck) {
+		this.deck = deck;
+	}
+
+	public String getName() {
+		return name;
+	}
+
+	public void setName(String name) {
+		this.name = name;
+	}
+
+	public Player(String name) {
+		this.setName(name);
+	}
+	
+}
