+++ olympics/src/main/java/org/gardler/biglittlechallenge/olympics/model/Player.java
@@ -1,28 +1,9 @@
 package org.gardler.biglittlechallenge.olympics.model;
 
-public class Player {
-
-	String name;
-	Deck deck;
-	
-	public Deck getDeck() {
-		return deck;
-	}
-
-	public void setDeck(Deck deck) {
-		this.deck = deck;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
+public class Player extends org.gardler.biglittlechallenge.core.model.Player {
 
 	public Player(String name) {
-		this.setName(name);
+		super(name);
 	}
-	
+
 }
